Implement created, update and periodic agenda items sync

+++ agenda/data/src/main/java/com/gkp/agenda/data/workers/DeletedItemsWorker.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.gkp.agenda.domain.model.AgendaItemType
+import com.gkp.auth.domain.session.SessionStorage
 import com.gkp.core.database.dao.DeletedAgendaItemsDao
 import com.gkp.core.network.TaskyRetrofitApi
 import kotlinx.coroutines.currentCoroutineContext
@@ -12,17 +13,23 @@ import kotlinx.coroutines.ensureActive
 class DeletedItemsWorker(
     private val taskyRetrofitApi: TaskyRetrofitApi,
     private val deletedAgendaItemsDao: DeletedAgendaItemsDao,
+    private val sessionStorage: SessionStorage,
     context: Context,
     workerParams: WorkerParameters,
 ) : CoroutineWorker(context, workerParams) {
 
     override suspend fun doWork(): Result {
+
+        if (sessionStorage.getAuthInfo().accessToken.isBlank()){
+            return Result.failure()
+        }
+        if (runAttemptCount >= 5) {
+            return Result.failure()
+        }
+
         val agendaItemId = inputData.getString(AGENDA_ITEM_ID) ?: return Result.failure()
         val agendaItemType = inputData.getString(AGENDA_ITEM_TYPE) ?: return Result.failure()
         val type = AgendaItemType.valueOf(agendaItemType)
-       if(runAttemptCount >= 5 ){
-           return Result.failure()
-       }
        return try {
            performDelete(agendaItemId, type)
            deletedAgendaItemsDao.deleteForId(agendaItemId)
